requirements.txt
Flask==3.0.0
Flask-MySQLdb==2.0.0
python-dotenv==1.0.0
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0
gunicorn==21.2.0
pytest==7.4.3
